Allow the username prefix to be empty
Some people use an authentication plugin, ultimately making the prefix useless. It's not recommended to use no prefix, but if authentication is handled by another plugin, then there is not really any harm in using it, and it opens up the opportunity for the same account to be used on bedrock and java.