change username-prefix from a non-win allowed character file name to allowed
Change the "*" (which is not allowed by default on any windows File System) to "°"a win1252 allowed character that looks similar (not star bit circle).
Why? because we from https://github.com/SkinsRestorer/SkinsRestorerX and many other plugins use flatfiles that do not co-operate to well with this.
If there are any questions please let us know.