Patch for people who updated to 2.2 early, and fixed previous push.
The config updater didn't re-add the quotes that strings with special characters (or no characters) should have, which resulted in the usernamePrefix becoming invalid or null. While this has been fixed for people that use the config updater with the latest version, people that already updated can't update again, and would have to edit the config manually.